% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blotter.R
\name{blotter}
\alias{blotter}
\title{Trade Blotter}
\usage{
blotter(obj)
}
\arguments{
\item{obj}{An R object}
}
\value{
If \code{obj} is an acceptable blotter, returns \code{obj}; \code{FALSE} otherwise.
}
\description{
The \code{blotter}() function accepts any R object \code{obj}. If \code{obj} is a blotter,
then \code{blotter}() returns \code{obj}. If not, then \code{blotter}() will return \code{FALSE}
and print information about which criteria were not satisfied.
}
\details{
\strong{Options exercises} do not appear explicitly on the blotter. However, the
purchase/sale of the option, and the purchase/sale of the underlying asset
due to an exercise, do.

\strong{Shorting} also does not appear explicitly on the blotter -- it appears
only as a sale.

\strong{Dividends, coupon payments, and other cash} do not appear on the blotter.
Trades only.
}
\section{Blotters}{


A \strong{blotter} is a time-series \link[xts]{xts} object that contains a
complete record of all trades made during a given time interval. To be a
blotter, an object must satisfy the following criteria:

\itemize{
\item{\strong{class}}: \code{obj} must be an xts.
\item{\strong{dimensions}}: \code{nrow(obj)} must have 1 or more rows.
\item{\strong{rows}}: Each row represents a completed (filled) trade
order for a single asset.
\item{\strong{columns}}: An xts object must have five columns: \emph{symbol},
\emph{action}, \emph{price}, \emph{size}, and \emph{fees}. No other names are accepted.
Case-sensitive.
\item{No missing values}: Every element in the frame must be filled.
}

A blotter's columns must meet the following criteria:

\itemize{
\item{\strong{index}}: The xts datetime index gives the timestamp (or
date) on which the trade was filled.
\item{\strong{symbol}}: Identifier for the asset traded. Can be a ticker,
CUSIP, ISIN, a unique ID assigned by a brokerage, etc.
\item{\strong{action}}: "BUY" or "SELL".
\item{\strong{price}}: Always positive. The asset's price for the trade.
\item{\strong{size}}: Number of shares, contracts, etc that were traded.
\item{\strong{fees}}: Sum of all fees paid for the trade. Includes
commission, transaction fees, etc. Does not include taxes.
}
}

