% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_daily_returns.R
\name{calculate_daily_returns}
\alias{calculate_daily_returns}
\title{Calculate Daily Returns}
\usage{
calculate_daily_returns(
  assets,
  date_range_xts = NULL,
  buy_at = c("close", "open", "high", "low", "price"),
  sell_at = c("close", "open", "high", "low", "price"),
  include_volume = TRUE,
  include_buy_sell = FALSE,
  returns_method = c("ln", "log2", "log10", "pct_diff", "multiple")
)
}
\arguments{
\item{assets}{a list of sub-lists named and corresponding to a given
asset, for example, "AAPL". Each sublist contains split, dividend, and
aggregated OHLCV data as xts objects. An example of an object that may be
passed as \code{asset_data} is \link{sample_historical_data}.}

\item{date_range_xts}{A string that specifies the date range for which
\code{calculate_daily_returns} will calculate returns. Gets passed to an
xts, so simply use xts subsetting notation. If missing or NULL, then the
full date range of OHLCV is used.}

\item{buy_at}{Character: Specifies the 'initial price' to be used for
calculating returns. Must be One of "open", "high", "low", "close", or
"price". Case sensitive.}

\item{sell_at}{Character: Specifies the 'final price' to be used for
calculating returns. Must be One of "open", "high", "low", "close", or
"price". Case sensitive.}

\item{include_volume}{Boolean: include volume data in output? Default TRUE.}

\item{include_buy_sell}{Boolean: include in the output the buy & sell prices
that were used to calculate each return? Default FALSE.}

\item{returns_method}{Character vector specifying how the returns should be
calculated. Choices are: "ln", "log2", "log10", "pct_diff", and "multiple".
Selecting multiple choices is allowed. Defaults to all.}
}
\value{
depends on value passed to \code{include_volume}:
\itemize{
\item \strong{DEFAULT}, \code{include_volume == FALSE}: returns an xts
object. Each element is the daily log return -- taking into account splits
and dividends -- of the asset specified in the column name, realized on
the date specified in the xts index relative to the previous trading day.
\item \code{include_volume == TRUE}: returns a list of two xts objects
named \code{"daily_returns"} and \code{"daily_volume"}.
\code{"daily_returns"} is exactly the same xts object that is returned if
\code{include_volume == TRUE}, as described above. \code{"daily_volume"}
has the same structure, date indices, and colnames \code{"daily_returns"},
but contains the trade volume observed for a given asset on a given day.
}
}
\description{
Calculate day-over-day historical returns for a list of assets, taking into
account splits and dividends.
}
