% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_of_rebalancing.R
\name{cost_of_rebalancing}
\alias{cost_of_rebalancing}
\title{Takes in two portfolios & current prices --> returns the cost of going from
portfolio 1 to portfolio 2 at those prices.}
\usage{
cost_of_rebalancing(portfolio_1, portfolio_2, prices)
}
\description{
Takes in two portfolios & current prices --> returns the cost of going from
portfolio 1 to portfolio 2 at those prices.
}
