% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{currency_identifiers}
\alias{currency_identifiers}
\title{Currency Identifiers}
\format{An tibble containing the following self-explanatory columns:
\code{three_letter_code}, \code{currency_name}, \code{symbol},
\code{country}}
\usage{
currency_identifiers
}
\description{
Names, symbols, and country of origin for currencies downloaded from
\url{https://www.xe.com/symbols.php}
}
\keyword{datasets}
