% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exchange_hours}
\alias{exchange_hours}
\title{Exchange Hours}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 66 rows and 13 columns.}
\usage{
exchange_hours
}
\description{
Operating hours and related information for worldwide exchanges. Taken from
\url{https://en.wikipedia.org/wiki/List_of_stock_exchange_trading_hours}.
}
\keyword{datasets}
