% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FinancieR-package.R
\docType{package}
\name{FinancieR-package}
\alias{FinancieR}
\alias{FinancieR-package}
\title{FinancieR: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Jake Vestal \email{jmvestal@gmail.com} (0000-0002-2970-4795)

}
\keyword{internal}
