% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inject_cash.R
\name{inject_cash}
\alias{inject_cash}
\title{Takes in a portfolio object & prices and calculates the new portfolio if
you injected cash ONLY.}
\usage{
inject_cash(
  portfolio,
  prices,
  exp_rtn,
  exp_vol,
  exp_cor,
  rfr = 2.7397e-05,
  portfolio_aum
)
}
\description{
Takes in a portfolio object & prices and calculates the new portfolio if
you injected cash ONLY.
}
