% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebalance_portfolio.R
\name{rebalance_portfolio}
\alias{rebalance_portfolio}
\title{Rebalance an Existing Portfolio}
\usage{
rebalance_portfolio(portfolio, exp_rtn, exp_vol, exp_cor, prices)
}
\arguments{
\item{exp_rtn}{Named numeric vector for which each element is the return
expected for the next period for the asset specified by the element's name.}

\item{exp_vol}{Named numeric vector for which each element is the volatility
expected for the asset specified by the element's name.}

\item{exp_cor}{Named numeric matrix specifying the expected covariance of
returns for each asset pair.}

\item{prices}{Optional: a named numeric vector for which each name is the
identifier of an asset and each element is the current price of that asset
for which a market portfolio is to be calculated on a shares basis. See the
"Returns" section for more info.}
}
\description{
Given an existing portfolio and a set of expected returns, volatilities, and
a correlation matrix, calculate the trades required and the MP that would
result if the portfolio were rebalanced at specified market prices. Assets
passed in via the parameters \emph{exp_rtn}, \emph{exp_vol}, and
\emph{exp_cor} specify the universe, or choices available, for inclusion in
the market portfolio, but not every asset passed in will necessarily appear
in the optimized MP.
}
\examples{
# Calculate the Sharpe-optimized market portfolio (MP) found for the stocks in 
# stock_data if built on 17 Jan 2019. Assume:
#   1) The return we expect over the next year for each asset is the annualized
#        GMMR of the daily rates of return over the past year.
#   2) The volatilities we expect over the next year for each asset are the
#        annualized daily vols we observed during the previous year.
#   3) The correlations of returns over the next year will be the same as they 
#        were for last year
#   4) Each asset may be acquired at the closing price on 17 Jan 2019.
#   5) We have $500,000 to invest.

historical_rtn <- calculate_historical_returns(
  assets         = stock_data,
  date_range_xts = paste0(as.Date("2019-01-17") - 365, "/2019-01-17")
)

portfolio_2019_01_17 <- calculate_market_portfolio(
  exp_rtn       = log(1 + gmrr(historical_rtn))*252,
  exp_vol       = (
    dplyr::summarize(
      tibble::as_tibble(historical_rtn),
      dplyr::across(dplyr::everything(), sd, na.rm = TRUE)
    ) \%>\%
      purrr::as_vector()
  ) * sqrt(252),
  exp_cor       = stats::cor(historical_rtn, use = "pairwise.complete.obs"),
  prices        = stock_data \%>\%
    lapply(function(stock){stock$prices$Close["2019-01-17"]}) \%>\%
    unlist() \%>\%
    purrr::compact(),
  portfolio_aum = 500000
)

}
