% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_info.R
\name{stock_info}
\alias{stock_info}
\title{Stock Info}
\usage{
stock_info(stocks = stock_data, symbols = names(stocks))
}
\arguments{
\item{stocks}{List of stocks following the same structure as
\link{stock_data}. Defaults to \emph{stock_data}.}

\item{symbols}{Character vector containing symbols to be included in the
output. Each element of \emph{symbols} must appear as a name of a stock
element in \emph{stocks}. Defaults to \emph{names(stocks)}.}
}
\value{
A \link[tibble]{tibble} containing available descriptive data for
each member of \code{stock_data}. Not case-sensitive.
}
\description{
Get all available information about a stock printed cleanly as a tibble.
\emph{stock_info}() is not a data object; rather, it's a function that
scans a data object (\emph{stock_data}) and compiles and returns the gathered
information in a readable format.
}
\section{Disclaimer}{
 Absolutely no promise, guarantee, or representation is
made by the authors of FinancieR concerning the accuracy or reliability of
\emph{stock_data}, or any other data in the FinancieR package.
}

\examples{
  stock_info()

}
