% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_stock.R
\name{[.stock}
\alias{[.stock}
\alias{extract_stock}
\title{Using \code{[} on Objects of Class "stock"}
\usage{
\method{[}{stock}(x, i, j = NULL, ..., silent = FALSE)
}
\arguments{
\item{x}{an \code{\link[xts]{xts}} subsetting string following the same rules
as for an \code{\link[xts]{xts}} object: "\strong{2015}" will return data
for all of 2015, "\strong{2018-04}" will return data for March of 2018
only, "\strong{2016-09-19/2016-12-31}" will return data from 19 Sep 2016 to
the end of the year, etc. If NULL, then the full time range of available
data for the selected stock will be returned.}

\item{j}{Character vector specifying column names to be included in the output. Available column names are:

"DividendAmount", "RecordDate", "DeclaredDate", "Type", "PaymentFrequency", "Open", "High", "Low", "Close", "TWAP", "VWAP", "Volume", "PreviousCloseDate", "PreviousClose", "Denominator", "Numerator", "ExDate"}

\item{silent}{Boolean. If FALSE (Default), messages will be displayed for
special cases that might produce unexpected behavior; e.g., if no data
is available for the date range specified. If TRUE, these messages will
not be displayed.}
}
\value{
an xts object, or NULL if no data is available for the specified
asset within the specified date range.
}
\description{
You can quickly retrieve price, dividend, split, and MnA data useing the \code{[}
operator on an object of class \emph{stock}, much like subsetting a matrix
or list in base R.
}
\details{
The rules are:
\enumerate{
\item If a dividend took place during the specified date range, then the
column "\strong{DividendAmount}" will always be included in the output.
\item If a split took place during the specified date range, then the
columns "\strong{Denominator}" and "\strong{Numerator}" will always be
included in the output. "\strong{Denominator}" refers to the number of
shares AFTER the split and "\strong{Numerator}" number of shares BEFORE
the split; for example, a 2-to-1 split would have "\strong{Denominator}"
= 2 and "\strong{Numerator}" = 1.
\item If an M&A event took place during the specified time range resulting
in a change in the stock's ticker symbol, then the columns
"\strong{symbol}" and "\strong{multiple}" will always be included in the
output. "\strong{symbol}" is the ticker symbol that the company traded
under through time, and "\strong{multiple}" gives the ratio
\emph{number of shares held prior to MnA} /
\emph{number of shares held after the MnA}.
}
}
\examples{

# If no columns are specified, all available columns are returned.
# Single date:
stock_data$IVV["2019-06-11"]


# Dividends are correctly applied, even when issued on a non-trading day. In
# 2012, for example, Texas Instruments (TXN) had scheduled a dividend for the
# 29th of October but the exchange was closed due to Hurricane Sandy. The
# dividend, then, should be applied to the next trading day (31 October). Here,
# we'll select just the closing price for a date range around that time.
# Remember, any dividends that occur within the date range are always included.
stock_data$TXN["2012-10-25/2012-11-05", "Close"]


# Mergers are handled as illustrated below for the PX-LIN merger in 2018.
# Shareholders of PX received 1 share of the new company, LIN, for every 1
# share of PX.
stock_data$PX["2018-10-29/2018-11-05", "Close"]

# Splits are handled as follows for Apple's 7-to-1 split of 2014:
stock_data$AAPL["2014-06-05/2014-06-11", c("Open", "Close")]

# Not specifying which columns you want can quickly lead to a lot of data, as in
# the below subset of IVV for the month of June 2019:
stock_data$IVV["2019-06"]
}
