% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trading_dates.R
\name{trading_dates}
\alias{trading_dates}
\title{Trading Dates}
\format{
date vector, ordered from oldest to most recent.
}
\usage{
trading_dates()
}
\description{
Returns a date vector containing all the dates for which a historical stock
price was reported by \strong{any} stock in \emph{stock_data}.
}
