% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yahoo_adj_prices}
\alias{yahoo_adj_prices}
\title{Yahoo Adjusted Prices}
\format{
An object of class \code{xts} (inherits from \code{zoo}) with 1259 rows and 11 columns.
}
\usage{
yahoo_adj_prices
}
\description{
Historical adjusted closing prices for 10 stock and an S&P500 tracking index
taken from Yahoo Finance. To be used for development and testing purposes
only -- data are not trading-quality.
}
\examples{
yahoo_adj_prices

}
\keyword{datasets}
