% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_stock.R
\name{as_stock}
\alias{as_stock}
\title{Convert to Class "\strong{\code{stock}}"}
\usage{
as_stock()
}
\description{
Coerce a list containing historical financial data for a stock, a stock's
options chain, or a commodity's price forward curve into an R object having
class \strong{asset}.
}
