% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_market_portfolio.R
\name{calculate_market_portfolio}
\alias{calculate_market_portfolio}
\title{Calculate Market Portfolio}
\usage{
calculate_market_portfolio(
  exp_rtn,
  exp_vol,
  exp_cor,
  rfr = 2.7397e-05,
  allow_shorts = FALSE,
  prices = NULL,
  portfolio_aum = NULL
)
}
\arguments{
\item{exp_rtn}{Named numeric vector for which each element is the return
expected for the asset specified by the element's name.}

\item{exp_vol}{Named numeric vector for which each element is the volatility
expected for the asset specified by the element's name.}

\item{exp_cor}{Named numeric matrix specifying the expected covariance of
returns for each asset pair.}

\item{rfr}{The risk-free rate (in decimal form; i.e., to specify a rate of
"3\%" use "0.03", \emph{not} "3") that \strong{YOU} can earn on cash with
reasonable liquidity constraints that \strong{YOU} can tolerate. For big
banks and in economic textbooks, this rate is usually the current rate on
3-month T-bills (or even higher). That might be just fine for you if you
plan on rebalancing once a year. For a trader running a strategy that can't
tolerate cash being tied up for 3 months at a time, the \emph{rfr} should
be set to whatever interest rate your brokerage is giving you on cash in
your trading account. Used in calculation of the
\href{https://www.investopedia.com/articles/07/sharpe_ratio.asp}{Sharpe
Ratio}. Defaults to \strong{0.0027397}\% daily return (about 1\% annually).
If you specify a different value for \emph{rfr}, \strong{make sure its time
basis matches the one used for you other inputs}, (i.e., if \emph{exp_rtn}
contains monthly returns, use monthly risk-free rate)!}

\item{allow_shorts}{Defaults to FALSE; set to TRUE to allow shorting of all
the assets. There are MANY ways to do this, but by default
\emph{calculate_market_portfolio}() simply treats shorts as another asset
whose expected return equals negative the expected return of going long.}

\item{prices}{Optional: a named numeric vector for which each name is the
identifier of an asset and each element is the current price of that asset
for which a market portfolio is to be calculated on a shares basis. See the
"Returns" section for more info.}

\item{portfolio_aum}{Optional: numeric, length 1, giving the total amount of
assets under management for which a market portfolio is to be calculated on
a shares basis. See "Returns" section for more info.}
}
\value{
If \strong{weights basis} (\emph{prices} and \emph{portfolio_aum}
not specified): #'   A list describing the market portfolio (MP). Contains
the following four elements:

\describe{
\item{\code{sharpe}, numeric:}{Sharpe Ratio of MP}
\item{\code{weights}, named numeric vector:}{The values of
\code{weights} range from 0 to 1 and denote the fraction of the total
MP value allocated to the asset whose identifier is that value's name.}
\item{\code{ex_return}, numeric:}{The return that is expected (i.e.,
predicted/forecast) for the MP over the next time interval.}
\item{\code{ex_volatility}:}{The volatility that is expected (i.e.,
predicted/forecast) for the MP over the next time interval.}
}

If \strong{shares basis} (\emph{prices} and \emph{portfolio_aum}
specified): \emph{calculate_market_portfolio}() will optimize the portfolio
assuming that  \emph{portfolio_aum} is available for investing and that the
price of each asset is given by \emph{prices}. Returns a list describing
the market portfolio (MP) which contains the for elements described above,
plus two more:

\describe{
\item{shares:}{Named numeric vector giving the number of shares of each
asset calculated for the market portfolio}
\item{cash:}{Numeric. Leftover cash not invested in assets.}
}
}
\description{
Calculate the Sharpe-optimal market portfolio available for a set of assets
given the expected returns, volatilities, and correlations of returns for
each asset.
}
\details{
All arguments which are percentages (\emph{exp_rtn}, \emph{exp_vol}, and
\emph{rfr}) must be supplied in decimal form; i.e., to specify "12\%", use
0.12, not 12.

It should go without saying, but make sure that every asset is represented
in the three inputs: names of \emph{exp_rtn}, \emph{exp_vol}, and the row &
colnames of \emph{exp_cor}.

This function works by finding the Sharpe-optimum equal-weighted portfolio
that can be created using the assets passed in. Using that portfolio as a
starting point, the function finds the assets A and B such that
reallocating ("selling") a small amount (\emph{step}) of A into asset B
("buying") results in a portfolio whose Sharpe is greater than all other
possible assets A and B. The portfolio variable is updated.

This process is repeated with smaller and smaller \emph{step} sizes. When
it is not possible to reallocate \emph{step} amount of any asset A into any
other asset B so as to create a portfolio having a better Sharpe, the
Market Portfolio has been reached and the function returns the value.
}
\examples{
# Let's pick an example date for which to create an MP:
mp_date <- "2018-01-09"
# Feel free to change this date to any other date for which you have data.

# Use calculate_returns() to get the daily returns observed for the 365 days
# ending on mp_date:
historical_rtn <- calculate_returns(
  assets         = stock_data,
  date_range_xts = paste0(
    as.Date(mp_date) - 365,
    "/",
    mp_date
  )
)

# We'll assume that the return we expect over the next year is the annualized 
# GMMR of the daily rates of return in historical_rtn:
exp_rtn <- gmrr(historical_rtn)

# Assume that the volatilities we expect for the next year are the annualized
# daily vols we observed during the previous year:
exp_vol <- dplyr::summarize(
  tibble::as_tibble(historical_rtn),
  dplyr::across(dplyr::everything(), sd, na.rm = TRUE)
) \%>\%
  purrr::as_vector()

# Assume that the correlations of returns of each asset pair that we expect
# for the next year will be the same as the previous year:
exp_cor <- stats::cor(historical_rtn, use = "pairwise.complete.obs")

# Calculate the market portfolio:
mp_by_wt <- calculate_market_portfolio(exp_rtn, exp_vol, exp_cor)
mp_by_wt

### Calculate the market portfolio allowing both long & short positions:
mp_by_wt_shorts <- calculate_market_portfolio(
  exp_rtn, exp_vol, exp_cor, allow_shorts = TRUE
)
mp_by_wt_shorts

### Repeat the above, assuming you have $250,000 to invest and that the stocks
### may be bought at their closing prices on mp_date:

prices <- stock_data \%>\%
  vapply(
    function(stock){
      stock$prices$Close[mp_date]
    },
    numeric(1)
  )
portfolio_aum <- 250000

mp_by_shares <- calculate_market_portfolio(
  exp_rtn,
  exp_vol,
  exp_cor,
  prices        = prices,
  portfolio_aum = portfolio_aum
)
mp_by_shares

### Allow shorting:
mp_by_shares_shorts <- calculate_market_portfolio(
  exp_rtn,
  exp_vol,
  exp_cor,
  allow_shorts  = TRUE,
  prices        = prices,
  portfolio_aum = portfolio_aum
)
mp_by_shares_shorts
}
