% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_returns.R
\name{calculate_returns}
\alias{calculate_returns}
\title{Calculate Returns}
\usage{
calculate_returns(
  assets,
  date_range_xts = NULL,
  buy_at = "Close",
  sell_at = "Close",
  returns_method = "ln"
)
}
\arguments{
\item{assets}{a list of sub-lists named and corresponding to a given
asset, for example, "AAPL". Each sublist contains split, dividend, and
aggregated prices data as xts objects. An example of an object that may be
passed as \emph{asset_data} is \link{stock_data}.}

\item{date_range_xts}{A string that specifies the date range for which
\emph{calculate_daily_returns}() will calculate returns. Use xts subsetting
notation to specify the date range; see examples in \link[xts]{xts}. If
missing or NULL, then the full date range of OHLCV prices is used.}

\item{buy_at}{Character: Specifies the 'initial price' to be used for
calculating returns. Must be one of "open", "high", "low", "close", or
"price", defaults to "Close". Not case sensitive.}

\item{sell_at}{Character: Specifies the 'final price' to be used for
calculating returns. Must be One of "open", "high", "low", "close", or
"price", defaults to "Close". Not case sensitive.}

\item{returns_method}{Character vector specifying how the returns should be
calculated, defaults to \strong{\emph{"ln}}. Choices are:
\itemize{
\item{\strong{"ln"}}: Log return; \emph{ln(sell_price / buy_price)} where
\emph{ln()} denotes the natural logarithm (base \emph{e}).
\item{\strong{"log2"}}: Log return; \emph{log(sell_price / buy_price)}
where \emph{log()} denotes the logarithm to base \emph{2}.
\item{\strong{"log10"}}: Log return; \emph{log(sell_price / buy_price)}
where \emph{log()} denotes the logarithm to base \emph{10}.
\item{\strong{"pct_diff"}}: Percent difference;
\emph{(sell_price - buy_price)/buy_price}. NOT multiplied by 100.
\item{\strong{"multiple"}}: Price multiple; \emph{sell_price / buy_price}.
}}
}
\value{
An xts object. Each element is the return observed on the date given
by the xts' index with respect to the previous period (taking into
account any splits and dividends that may have occurred) for the asset
specified by the element's column name.
}
\description{
Calculate period-over-period historical returns for a list of assets, taking
into account splits and dividends.
}
\examples{
# Calculate the daily natural log returns of Apple and AT&T for 2014, a year
# in which Apple had a stock split. Natural log returns, takes into account
# divs and splits.


}
