% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{currency_identifiers}
\alias{currency_identifiers}
\title{Currency Identifiers}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 109 rows and 4 columns.
}
\usage{
currency_identifiers
}
\description{
Names, symbols, and country of origin for currencies downloaded from \url{https://www.xe.com/symbols.php} stored in a tibble
}
\keyword{datasets}
