% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{currency_identifiers}
\alias{currency_identifiers}
\title{Currency Identifiers}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 109 rows and 4 columns.}
\usage{
currency_identifiers
}
\description{
Names, symbols, and country of origin for currencies downloaded from \url{https://www.xe.com/symbols.php} stored in a tibble: \tabular{llll}{ \strong{three_letter_code} \tab \strong{currency_name} \tab \strong{symbol} \tab \strong{country}          \cr  ALL                         \tab Lek                     \tab Lek              \tab Albania                    \cr  AFN                         \tab Afghani                 \tab ؋                \tab Afghanistan                \cr  ARS                         \tab Peso                    \tab $                \tab Argentina                  \cr  AWG                         \tab Guilder                 \tab ƒ                \tab Aruba                      \cr  AUD                         \tab Dollar                  \tab $                \tab Australia                  \cr  AZN                         \tab Manat                   \tab ₼                \tab Azerbaijan                 \cr  BSD                         \tab Dollar                  \tab $                \tab Bahamas                    \cr  BBD                         \tab Dollar                  \tab $                \tab Barbados                   \cr  BYN                         \tab Ruble                   \tab Br               \tab Belarus                    \cr  BZD                         \tab Dollar                  \tab BZ$              \tab Belize                     \cr  BMD                         \tab Dollar                  \tab $                \tab Bermuda                    \cr  BOB                         \tab Bolíviano               \tab $b               \tab Bolivia                    \cr  BAM                         \tab Convertible Marka       \tab KM               \tab Bosnia and Herzegovina     \cr  BWP                         \tab Pula                    \tab P                \tab Botswana                   \cr  BGN                         \tab Lev                     \tab лв               \tab Bulgaria                   \cr  BRL                         \tab Real                    \tab R$               \tab Brazil                     \cr  BND                         \tab Dollar                  \tab $                \tab Brunei Darussalam          \cr  KHR                         \tab Riel                    \tab ៛                \tab Cambodia                   \cr  CAD                         \tab Dollar                  \tab $                \tab Canada                     \cr  KYD                         \tab Dollar                  \tab $                \tab Cayman Islands             \cr  CLP                         \tab Peso                    \tab $                \tab Chile                      \cr  CNY                         \tab Yuan Renminbi           \tab ¥                \tab China                      \cr  COP                         \tab Peso                    \tab $                \tab Colombia                   \cr  CRC                         \tab Colon                   \tab ₡                \tab Costa Rica                 \cr  HRK                         \tab Kuna                    \tab kn               \tab Croatia                    \cr  CUP                         \tab Peso                    \tab ₱                \tab Cuba                       \cr  CZK                         \tab Koruna                  \tab Kč               \tab Czech Republic             \cr  DKK                         \tab Krone                   \tab kr               \tab Denmark                    \cr  DOP                         \tab Peso                    \tab RD$              \tab Dominican Republic         \cr  XCD                         \tab Dollar                  \tab $                \tab East Caribbean             \cr  EGP                         \tab Pound                   \tab £                \tab Egypt                      \cr  SVC                         \tab Colon                   \tab $                \tab El Salvador                \cr  EUR                         \tab Countries               \tab €                \tab Euro Member                \cr  FKP                         \tab Pound                   \tab £                \tab Falkland Islands (Malvinas)\cr  FJD                         \tab Dollar                  \tab $                \tab Fiji                       \cr  GHS                         \tab Cedi                    \tab ¢                \tab Ghana                      \cr  GIP                         \tab Pound                   \tab £                \tab Gibraltar                  \cr  GTQ                         \tab Quetzal                 \tab Q                \tab Guatemala                  \cr  GGP                         \tab Pound                   \tab £                \tab Guernsey                   \cr  GYD                         \tab Dollar                  \tab $                \tab Guyana                     \cr  HNL                         \tab Lempira                 \tab L                \tab Honduras                   \cr  HKD                         \tab Dollar                  \tab $                \tab Hong Kong                  \cr  HUF                         \tab Forint                  \tab Ft               \tab Hungary                    \cr  ISK                         \tab Krona                   \tab kr               \tab Iceland                    \cr  INR                         \tab Rupee                   \tab ₹                \tab India                      \cr  IDR                         \tab Rupiah                  \tab Rp               \tab Indonesia                  \cr  IRR                         \tab Rial                    \tab ﷼                \tab Iran                       \cr  IMP                         \tab Pound                   \tab £                \tab Isle of Man                \cr  ILS                         \tab Shekel                  \tab ₪                \tab Israel                     \cr  JMD                         \tab Dollar                  \tab J$               \tab Jamaica                    \cr  JPY                         \tab Yen                     \tab ¥                \tab Japan                      \cr  JEP                         \tab Pound                   \tab £                \tab Jersey                     \cr  KZT                         \tab Tenge                   \tab лв               \tab Kazakhstan                 \cr  KPW                         \tab Won                     \tab ₩                \tab Korea (North)              \cr  KRW                         \tab Won                     \tab ₩                \tab Korea (South)              \cr  KGS                         \tab Som                     \tab лв               \tab Kyrgyzstan                 \cr  LAK                         \tab Kip                     \tab ₭                \tab Laos                       \cr  LBP                         \tab Pound                   \tab £                \tab Lebanon                    \cr  LRD                         \tab Dollar                  \tab $                \tab Liberia                    \cr  MKD                         \tab Denar                   \tab ден              \tab Macedonia                  \cr  MYR                         \tab Ringgit                 \tab RM               \tab Malaysia                   \cr  MUR                         \tab Rupee                   \tab ₨                \tab Mauritius                  \cr  MXN                         \tab Peso                    \tab $                \tab Mexico                     \cr  MNT                         \tab Tughrik                 \tab ₮                \tab Mongolia                   \cr  MZN                         \tab Metical                 \tab MT               \tab Mozambique                 \cr  NAD                         \tab Dollar                  \tab $                \tab Namibia                    \cr  NPR                         \tab Rupee                   \tab ₨                \tab Nepal                      \cr  ANG                         \tab Guilder                 \tab ƒ                \tab Netherlands Antilles       \cr  NZD                         \tab Dollar                  \tab $                \tab New Zealand                \cr  NIO                         \tab Cordoba                 \tab C$               \tab Nicaragua                  \cr  NGN                         \tab Naira                   \tab ₦                \tab Nigeria                    \cr  NOK                         \tab Krone                   \tab kr               \tab Norway                     \cr  OMR                         \tab Rial                    \tab ﷼                \tab Oman                       \cr  PKR                         \tab Rupee                   \tab ₨                \tab Pakistan                   \cr  PAB                         \tab Balboa                  \tab B/.              \tab Panama                     \cr  PYG                         \tab Guarani                 \tab Gs               \tab Paraguay                   \cr  PEN                         \tab Sol                     \tab S/.              \tab Peru                       \cr  PHP                         \tab Peso                    \tab ₱                \tab Philippines                \cr  PLN                         \tab Zloty                   \tab zł               \tab Poland                     \cr  QAR                         \tab Riyal                   \tab ﷼                \tab Qatar                      \cr  RON                         \tab Leu                     \tab lei              \tab Romania                    \cr  RUB                         \tab Ruble                   \tab ₽                \tab Russia                     \cr  SHP                         \tab Pound                   \tab £                \tab Saint Helena               \cr  SAR                         \tab Riyal                   \tab ﷼                \tab Saudi Arabia               \cr  RSD                         \tab Dinar                   \tab Дин.             \tab Serbia                     \cr  SCR                         \tab Rupee                   \tab ₨                \tab Seychelles                 \cr  SGD                         \tab Dollar                  \tab $                \tab Singapore                  \cr  SBD                         \tab Dollar                  \tab $                \tab Solomon Islands            \cr  SOS                         \tab Shilling                \tab S                \tab Somalia                    \cr  ZAR                         \tab Rand                    \tab R                \tab South Africa               \cr  LKR                         \tab Rupee                   \tab ₨                \tab Sri Lanka                  \cr  SEK                         \tab Krona                   \tab kr               \tab Sweden                     \cr  CHF                         \tab Franc                   \tab CHF              \tab Switzerland                \cr  SRD                         \tab Dollar                  \tab $                \tab Suriname                   \cr  SYP                         \tab Pound                   \tab £                \tab Syria                      \cr  TWD                         \tab Dollar                  \tab NT$              \tab Taiwan New                 \cr  THB                         \tab Baht                    \tab ฿                \tab Thailand                   \cr  TTD                         \tab Dollar                  \tab TT$              \tab Trinidad and Tobago        \cr  TRY                         \tab Lira                    \tab ₺                \tab Turkey                     \cr  TVD                         \tab Dollar                  \tab $                \tab Tuvalu                     \cr  UAH                         \tab Hryvnia                 \tab ₴                \tab Ukraine                    \cr  GBP                         \tab Pound                   \tab £                \tab United Kingdom             \cr  USD                         \tab Dollar                  \tab $                \tab United States              \cr  UYU                         \tab Peso                    \tab $U               \tab Uruguay                    \cr  UZS                         \tab Som                     \tab лв               \tab Uzbekistan                 \cr  VEF                         \tab Bolívar                 \tab Bs               \tab Venezuela                  \cr  VND                         \tab Dong                    \tab ₫                \tab Viet Nam                   \cr  YER                         \tab Rial                    \tab ﷼                \tab Yemen                      \cr  ZWD                         \tab Dollar                  \tab Z$               \tab Zimbabwe                   }
}
\keyword{datasets}
