% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exchange_hours}
\alias{exchange_hours}
\title{Exchange Hours}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 66 rows and 13 columns.}
\usage{
exchange_hours
}
\description{
Operating hours and related information for worldwide exchanges. Downloaded from \url{https://www.xe.com/symbols.php} storedin a tibble: \tabular{lllllllllllll}{ \strong{Name}                            \tab \strong{ID} \tab \strong{Country}                \tab \strong{City}  \tab \strong{Zone} \tab \strong{Δ} \tab \strong{DST} \tab \strong{local_Open} \tab \strong{local_Close} \tab \strong{local_Lunch}                   \tab \strong{UTC_Open} \tab \strong{UTC_Close} \tab \strong{UTC_Lunch}                    \cr  New Zealand Stock Market                  \tab NZSX         \tab New Zealand                      \tab Wellington      \tab NZST           \tab +12         \tab Sep–Apr       \tab 10:00                \tab 16:45                 \tab No                                      \tab 22:00              \tab 05:00               \tab No                                     \cr  Australian Securities Exchange            \tab ASX          \tab Australia                        \tab Sydney          \tab AEST           \tab +10         \tab Oct–Apr       \tab 10:00                \tab 16:00                 \tab No                                      \tab 00:00              \tab 06:00               \tab No                                     \cr  Tokyo Stock Exchange                      \tab TSE          \tab Japan                            \tab Tokyo           \tab JST            \tab +9          \tab               \tab 09:00                \tab 15:00\link{2}              \tab 11:30–12:30                             \tab 00:00              \tab 06:00               \tab 02:30–03:30                            \cr  Korea Stock Exchange                      \tab KRX          \tab South Korea                      \tab Busan & Seoul   \tab KST            \tab +9          \tab               \tab 09:00                \tab 15:30                 \tab No                                      \tab 00:00              \tab 06:00               \tab No                                     \cr  Bursa Malaysia                            \tab MYX          \tab Malaysia                         \tab Kuala Lumpur    \tab MYT            \tab +8          \tab               \tab 09:00                \tab 17:00                 \tab 12:30–14:30                             \tab 01:00              \tab 09:00               \tab 04:30–06:30                            \cr  Singapore Exchange                        \tab SGX          \tab Singapore                        \tab Singapore       \tab SGT            \tab +8          \tab               \tab 09:00                \tab 17:00                 \tab 12:00–13:00\link{3}                          \tab 01:00              \tab 09:00               \tab Yes                                    \cr  Taiwan Stock Exchange                     \tab TWSE         \tab Taiwan (Republic of China)       \tab Taipei          \tab CST            \tab +8          \tab               \tab 09:00                \tab 13:30                 \tab No                                      \tab 01:00              \tab 05:30               \tab No                                     \cr  Hong Kong Futures Exchange                \tab HKFE         \tab Hong Kong                        \tab Hong Kong       \tab HKT            \tab +8          \tab               \tab 09:15                \tab 16:00                 \tab 12:00–13:00                             \tab 01:15              \tab 08:00               \tab 04:00–05:00                            \cr  Hong Kong Stock Exchange                  \tab HKEX         \tab Hong Kong                        \tab Hong Kong       \tab HKT            \tab +8          \tab               \tab 09:30\link{4}             \tab 16:00                 \tab 12:00–13:00                             \tab 01:30              \tab 08:00               \tab 04:00–05:00                            \cr  Shanghai Stock Exchange                   \tab SSE          \tab China                            \tab Shanghai        \tab CST            \tab +8          \tab               \tab 09:30                \tab 15:00\link{notes 4}        \tab 11:30–13:00                             \tab 01:30              \tab 07:00               \tab 03:30–05:00                            \cr  Shenzhen Stock Exchange                   \tab SZSE         \tab China                            \tab Shenzhen        \tab CST            \tab +8          \tab               \tab 09:30                \tab 15:00\link{notes 4}        \tab 11:30–13:00                             \tab 01:30              \tab 07:00               \tab 03:30–05:00                            \cr  Saint Vincent and the Grenadines Exchange \tab SVGEX        \tab Saint Vincent and the Grenadines \tab Kingstown       \tab AST            \tab - 4         \tab               \tab 09:00                \tab 16:00\link{notes 4}        \tab 11:30–13:00                             \tab 13:00              \tab 20:00               \tab 15:30–17:00                            \cr  Philippine Stock Exchange                 \tab PSE          \tab Philippines                      \tab Manila          \tab PHT            \tab +8          \tab               \tab 09:30                \tab 15:30                 \tab 12:00–13:30                             \tab 01:30              \tab 07:30               \tab 04:00–05:30                            \cr  Indonesia Stock Exchange                  \tab IDX          \tab Indonesia                        \tab Jakarta         \tab WIB            \tab +7          \tab               \tab 09:00                \tab 16:00                 \tab 12:00-13:30 (Mon-Thu) 11:30-14:00 (Fri) \tab 02:00              \tab 09:00               \tab 05:00-06:30 (Mon-Thu) 04:30-07:00 (Fri)\cr  Hochiminh Stock Exchange                  \tab HOSE         \tab Vietnam                          \tab Hochiminh       \tab ICT            \tab +7          \tab               \tab 09:00                \tab 14:45                 \tab 11:30–13:00                             \tab 02:00              \tab 07:45               \tab 04:30–06:00                            \cr  Hanoi Stock Exchange                      \tab HNX          \tab Vietnam                          \tab Hanoi           \tab ICT            \tab +7          \tab               \tab 09:00                \tab 14:45                 \tab 11:30–13:00                             \tab 02:00              \tab 07:45               \tab 04:30–06:00                            \cr  Stock Exchange of Thailand                \tab SET          \tab Thailand                         \tab Bangkok         \tab ICT            \tab +7          \tab               \tab 10:00                \tab 16:30                 \tab 12:30–14:30                             \tab 03:00              \tab 09:30               \tab 05:30–07:30                            \cr  Chittagong Stock Exchange                 \tab CSE          \tab Bangladesh                       \tab Chittagong      \tab BST            \tab +6          \tab               \tab 10:30                \tab 14:30                 \tab No                                      \tab 04:30              \tab 08:30               \tab No                                     \cr  Dhaka Stock Exchange                      \tab DSE          \tab Bangladesh                       \tab Dhaka           \tab BST            \tab +6          \tab               \tab 10:30                \tab 14:30                 \tab No                                      \tab 04:30              \tab 08:30               \tab No                                     \cr  Bombay Stock Exchange\link{5}                  \tab BSE          \tab India                            \tab Mumbai          \tab IST            \tab +5.5        \tab               \tab 09:15                \tab 15:30                 \tab No                                      \tab 03:45              \tab 10:00               \tab No                                     \cr  National Stock Exchange of India\link{6}       \tab NSE          \tab India                            \tab Mumbai          \tab IST            \tab +5.5        \tab               \tab 09:15                \tab 15:30                 \tab No                                      \tab 03:45              \tab 10:00               \tab No                                     \cr  Colombo Stock Exchange                    \tab CSE          \tab Sri Lanka                        \tab Colombo         \tab SLST           \tab +5.5        \tab               \tab 09:30                \tab 14:30                 \tab No                                      \tab 04:00              \tab 09:00               \tab No                                     \cr  Pakistan Stock Exchange                   \tab PSX          \tab Pakistan                         \tab Karachi         \tab PKT            \tab +5          \tab               \tab 09:30                \tab 15:30                 \tab No                                      \tab 04:30              \tab 10:30               \tab No                                     \cr  Tehran Stock Exchange                     \tab TSE          \tab Iran                             \tab Tehran          \tab IRST           \tab +3.5        \tab               \tab 09:00                \tab 12:30                 \tab No                                      \tab 05:30              \tab 09:00               \tab No                                     \cr  Dar es Salaam Stock Exchange              \tab DSE          \tab Tanzania                         \tab Dar es Salaam   \tab EAT            \tab +3          \tab               \tab 08:00                \tab 17:00                 \tab No                                      \tab 05:00              \tab 14:00               \tab No                                     \cr  Moscow Exchange                           \tab MOEX         \tab Russia                           \tab Moscow          \tab MSK            \tab +3          \tab               \tab 10:00                \tab 18:45                 \tab No                                      \tab 07:00              \tab 15:45               \tab No                                     \cr  Saudi Stock Exchange                      \tab TADAWUL      \tab Saudi Arabia                     \tab Riyadh          \tab AST            \tab +3          \tab               \tab 10:00                \tab 15:00                 \tab No                                      \tab 07:00              \tab 12:00               \tab No                                     \cr  Nairobi Securities Exchange               \tab NSE          \tab Kenya                            \tab Nairobi         \tab UTC            \tab +3          \tab               \tab 09:30                \tab 15:00                 \tab No                                      \tab 06:30              \tab 12:00               \tab No                                     \cr  Helsinki Stock Exchange                   \tab OMX          \tab Finland                          \tab Helsinki        \tab EET            \tab +2          \tab Mar–Oct       \tab 10:00                \tab 18:30                 \tab No                                      \tab 08:00              \tab 16:30               \tab No                                     \cr  Ukrainian Exchange                        \tab UX           \tab Ukraine                          \tab Kiev            \tab EET            \tab +2          \tab Mar–Oct       \tab 10:00                \tab 17:30                 \tab No                                      \tab 08:00              \tab 15:30               \tab No                                     \cr  Riga Stock Exchange                       \tab OMXR         \tab Latvia                           \tab Riga            \tab EET            \tab +2          \tab Mar–Oct       \tab 10:00                \tab 16:00                 \tab No                                      \tab 08:00              \tab 14:00               \tab No                                     \cr  Tallinn Stock Exchange                    \tab OMXT         \tab Estonia                          \tab Tallinn         \tab EET            \tab +2          \tab Mar–Oct       \tab 10:00                \tab 16:00                 \tab No                                      \tab 08:00              \tab 14:00               \tab No                                     \cr  NASDAQ OMX Vilnius                        \tab OMXV         \tab Lithuania                        \tab Vilnius         \tab EET            \tab +2          \tab Mar–Oct       \tab 10:00                \tab 16:00                 \tab No                                      \tab 08:00              \tab 14:00               \tab No                                     \cr  Amman Stock Exchange                      \tab ASE          \tab Jordan                           \tab Amman           \tab EET            \tab +2          \tab Mar–Oct       \tab 10:00                \tab 12:00                 \tab No                                      \tab 08:00              \tab 10:00               \tab No                                     \cr  Tel Aviv Stock Exchange                   \tab TASE         \tab Israel                           \tab Tel Aviv        \tab IST            \tab +2          \tab Mar–Oct       \tab 09:00                \tab 17:30                 \tab No                                      \tab 07:00              \tab 15:30               \tab No                                     \cr  Beirut Stock Exchange                     \tab BSE          \tab Lebanon                          \tab Beirut          \tab EET            \tab +2          \tab Mar–Oct       \tab 09:30                \tab 12:30                 \tab No                                      \tab 07:30              \tab 10:30               \tab No                                     \cr  Johannesburg Stock Exchange               \tab JSE          \tab South Africa                     \tab Johannesburg    \tab SAST           \tab +2          \tab               \tab 09:00                \tab 17:00                 \tab No                                      \tab 07:00              \tab 15:00               \tab No                                     \cr  Borsa Istanbul                            \tab ISE          \tab Turkey                           \tab Istanbul        \tab TRT            \tab +3          \tab               \tab 10:00                \tab 18:00\link[=8]{7}           \tab 13:00–14:00                             \tab 09:00              \tab 16:00               \tab 11:00–12:00                            \cr  Frankfurt Stock Exchange (Xetra)          \tab FSX          \tab Germany                          \tab Frankfurt       \tab CET            \tab +1          \tab Mar–Oct       \tab 08:00                \tab 20:00\link{notes 5}        \tab No                                      \tab 07:00              \tab 19:00               \tab No                                     \cr  Eurex Exchange                            \tab EUREX        \tab Germany                          \tab Eschborn        \tab CET            \tab +1          \tab Mar–Oct       \tab 08:00                \tab 22:00                 \tab No                                      \tab 07:00              \tab 21:00               \tab No                                     \cr  Warsaw Stock Exchange                     \tab GPW          \tab Poland                           \tab Warsaw          \tab CET            \tab +1          \tab Mar–Oct       \tab 09:00                \tab 17:00\link{9}              \tab No                                      \tab 08:00              \tab 16:00               \tab No                                     \cr  Wiener Börse AG                           \tab VSE          \tab Austria                          \tab Vienna          \tab CET            \tab +1          \tab Mar–Oct       \tab 08:55                \tab 17:35                 \tab No                                      \tab 07:55              \tab 16:35               \tab No                                     \cr  Budapest Stock Exchange                   \tab BSE          \tab Hungary                          \tab Budapest        \tab CET            \tab +1          \tab Mar–Oct       \tab 09:00                \tab 17:00\link{10}             \tab No                                      \tab 08:00              \tab 16:00               \tab No                                     \cr  Euronext Paris                            \tab EPA          \tab France                           \tab Paris           \tab CET            \tab +1          \tab Mar–Oct       \tab 09:00                \tab 17:30                 \tab No                                      \tab 08:00              \tab 16:30               \tab No                                     \cr  Casablanca Stock Exchange                 \tab CSE          \tab Maroc                            \tab Casablanca      \tab CET            \tab +0          \tab Mar–Oct       \tab 08:10                \tab 15:55                 \tab No                                      \tab 08:10              \tab 15:55               \tab No                                     \cr  Swiss Exchange                            \tab SIX          \tab Switzerland                      \tab Zurich          \tab CET            \tab +1          \tab Mar–Oct       \tab 09:00                \tab 17:30                 \tab No                                      \tab 08:00              \tab 16:30               \tab No                                     \cr  Berne eXchange                            \tab BX           \tab Switzerland                      \tab Berne           \tab CET            \tab +1          \tab Mar–Oct       \tab 09:00                \tab 16:30                 \tab No                                      \tab 08:00              \tab 15:30               \tab No                                     \cr  Spanish Stock Exchange                    \tab BME          \tab Spain                            \tab Madrid          \tab CET            \tab +1          \tab Mar–Oct       \tab 09:00                \tab 17:30                 \tab No                                      \tab 08:00              \tab 16:30               \tab No                                     \cr  Milan Stock Exchange                      \tab MTA          \tab Italy                            \tab Milan           \tab CET            \tab +1          \tab Mar–Oct       \tab 09:00                \tab 17:35                 \tab No                                      \tab 08:00              \tab 16:35               \tab No                                     \cr  Euronext Amsterdam                        \tab AMS          \tab Netherlands                      \tab Amsterdam       \tab CET            \tab +1          \tab Mar–Oct       \tab 09:00                \tab 17:40                 \tab No                                      \tab 08:00              \tab 16:40               \tab No                                     \cr  Luxembourg Stock Exchange                 \tab LuxSE        \tab Luxembourg                       \tab Luxembourg City \tab CET            \tab +1          \tab Mar–Oct       \tab 09:00                \tab 17:35                 \tab No                                      \tab 08:00              \tab 16:35               \tab No                                     \cr  Stockholm Stock Exchange                  \tab OMX          \tab Sweden                           \tab Stockholm       \tab CET            \tab +1          \tab Mar–Oct       \tab 09:00                \tab 17:30                 \tab No                                      \tab 08:00              \tab 16:30               \tab No                                     \cr  Oslo Stock Exchange                       \tab OSE          \tab Norway                           \tab Oslo            \tab CET            \tab +1          \tab Mar–Oct       \tab 09:00                \tab 16:30                 \tab No                                      \tab 08:00              \tab 15:30               \tab No                                     \cr  Copenhagen Stock Exchange                 \tab CSE          \tab Denmark                          \tab Copenhagen      \tab CET            \tab +1          \tab Mar–Oct       \tab 09:00                \tab 17:00                 \tab No                                      \tab 08:00              \tab 16:00               \tab No                                     \cr  Malta Stock Exchange                      \tab MSE          \tab Malta                            \tab Valletta        \tab CET            \tab +1          \tab Mar–Oct       \tab 09:30                \tab 12:30\link{11}             \tab No                                      \tab 08:30              \tab 11:30               \tab No                                     \cr  Nigerian Stock Exchange                   \tab NSE          \tab Nigeria                          \tab Lagos           \tab WAT            \tab +1          \tab               \tab 10:00                \tab 16:00                 \tab No                                      \tab 09:00              \tab 15:00               \tab No                                     \cr  London Stock Exchange                     \tab LSE          \tab United Kingdom                   \tab London          \tab GMT            \tab +0          \tab Mar–Oct       \tab 08:00                \tab 16:30                 \tab No                                      \tab 08:00              \tab 16:30               \tab No                                     \cr  Euronext Lisbon                           \tab PSI          \tab Portugal                         \tab Lisbon          \tab GMT            \tab +0          \tab Mar–Oct       \tab 08:00                \tab 16:30                 \tab No                                      \tab 08:00              \tab 16:30               \tab No                                     \cr  Irish Stock Exchange                      \tab ISE          \tab Ireland                          \tab Dublin          \tab GMT            \tab +0          \tab Mar–Oct       \tab 08:00                \tab 16:30                 \tab No                                      \tab 08:00              \tab 16:30               \tab No                                     \cr  Bolsa de Valores de São Paulo             \tab Bovespa      \tab Brazil                           \tab São Paulo       \tab BRT            \tab −3          \tab Oct–Feb       \tab 10:00                \tab 17:30                 \tab No                                      \tab 13:00              \tab 20:00               \tab No                                     \cr  Buenos Aires Stock Exchange               \tab BCBA         \tab Argentina                        \tab Buenos Aires    \tab ART            \tab −3          \tab               \tab 11:00                \tab 17:00                 \tab No                                      \tab 14:00              \tab 20:00               \tab No                                     \cr  New York Stock Exchange                   \tab NYSE         \tab United States                    \tab New York        \tab EST            \tab −5          \tab Mar–Nov       \tab 09:30                \tab 16:00\link{12}             \tab No                                      \tab 14:30              \tab 21:00               \tab No                                     \cr  NASDAQ                                    \tab NASDAQ       \tab United States                    \tab New York        \tab EST            \tab −5          \tab Mar–Nov       \tab 09:30                \tab 16:00\link{13}             \tab No                                      \tab 14:30              \tab 21:00               \tab No                                     \cr  Toronto Stock Exchange                    \tab TSX          \tab Canada                           \tab Toronto         \tab EST            \tab −5          \tab Mar–Nov       \tab 09:30                \tab 16:00                 \tab No                                      \tab 14:30              \tab 21:00               \tab No                                     \cr  Mexican Stock Exchange                    \tab BMV          \tab Mexico                           \tab Mexico City     \tab CST            \tab −6          \tab Apr–Oct       \tab 08:30                \tab 15:00                 \tab No                                      \tab 14:30              \tab 21:00               \tab No                                     \cr  Jamaica Stock Exchange                    \tab JSE          \tab Jamaica                          \tab Kingston        \tab EST            \tab −5          \tab               \tab 09:00                \tab 13:00\link{14}             \tab No                                      \tab 14:00              \tab 17:00               \tab No                                     }
}
\keyword{datasets}
