% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FinancieR-package.R
\docType{package}
\name{FinancieR-package}
\alias{FinancieR}
\alias{FinancieR-package}
\title{FinancieR: Tools for Financial Data Analysis and Manipulation}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gothic-hedge-society/FinancieR}
  \item Report bugs at \url{https://github.com/gothic-hedge-society/FinancieR/issues}
}

}
\author{
\strong{Maintainer}: Jake Vestal \email{jmvestal@gmail.com} (\href{https://orcid.org/0000-0002-2970-4795}{ORCID})

}
\keyword{internal}
