% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_portfolio_from_hist_rtn.R
\name{market_portfolio_from_hist_rtn}
\alias{market_portfolio_from_hist_rtn}
\title{Market Portfolio (MP) from Historical Returns}
\usage{
market_portfolio_from_hist_rtn(returns_xts, rf = 0.0025)
}
\arguments{
\item{returns_xts}{An xts object whose values are period-over-period returns
observed for the assets specified by the identifiers in the column names.}

\item{rf}{The current prevailing risk-free rate.}
}
\value{
A list describing the market portfolio (MP) found by
\code{market_portfolio_from_hist_rtn()}, having four elements:
\describe{
\item{\code{sharpe}, numeric:}{Sharpe Ratio of MP}
\item{\code{weights}, named numeric vector:}{The values of
\code{weights} range from 0 to 1 and denote the fraction of the total
MP value allocated to the asset whose identifier is that value's name.}
\item{\code{ex_return}, numeric:}{The return that is expected (i.e.,
predicted/forecast) for the MP over the next time interval.}
\item{\code{ex_volatility}:}{The volatility that is expected (i.e.,
predicted/forecast) for the MP over the next time interval.}
}
}
\description{
Calculate optimal portfolio weighting fractions
}
\details{
Returns are NOT assumed to be in 'percent form': i.e., make sure that in
whatever xts is passed as \code{returns_xts}, a return of \emph{12\%} is
represented as \emph{0.12}.
}
