% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebalance_portfolio.R
\name{rebalance_portfolio}
\alias{rebalance_portfolio}
\title{Rebalance an Existing Portfolio}
\usage{
rebalance_portfolio(portfolio, exp_rtn, exp_vol, exp_cor, prices)
}
\arguments{
\item{exp_rtn}{Named numeric vector for which each element is the return
expected for the next period for the asset specified by the element's name.}

\item{exp_vol}{Named numeric vector for which each element is the volatility
expected for the asset specified by the element's name.}

\item{exp_cor}{Named numeric matrix specifying the expected covariance of
returns for each asset pair.}

\item{prices}{Optional: a named numeric vector for which each name is the
identifier of an asset and each element is the current price of that asset
for which a market portfolio is to be calculated on a shares basis. See the
"Returns" section for more info.}
}
\description{
Given an existing portfolio and a set of expected returns, volatilities, and
a correlation matrix, calculate the trades required and the MP that would
result if the portfolio were rebalanced at specified market prices. Assets
passed in via the parameters \emph{exp_rtn}, \emph{exp_vol}, and
\emph{exp_cor} specify the universe, or choices available, for inclusion in
the market portfolio, but not every asset passed in will necessarily appear
in the optimized MP.
}
