% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stock_exchange_hours}
\alias{stock_exchange_hours}
\title{Exchange Hours}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 66 rows and 13 columns.}
\usage{
stock_exchange_hours
}
\description{
Operating hours and related information for worldwide exchanges. Downloaded from \url{https://www.xe.com/symbols.php} storedin a tibble:
}
\keyword{datasets}
