% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withdraw_cash.R
\name{withdraw_cash}
\alias{withdraw_cash}
\title{Optimum way to withdraw cash.}
\usage{
withdraw_cash(portfolio, exp_rtn, exp_vol, exp_cov, prices, cash)
}
\description{
Taking into account capital gains taxes
}
