% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xts_merge_align_next.R
\name{xts_merge_align_next}
\alias{xts_merge_align_next}
\title{Merge 2 XTS Tables According to Next Non-Missing Value}
\usage{
xts_merge_align_next(xts1, xts2, agg_function, na.fill)
}
\arguments{
\item{xts1}{An XTS object}

\item{xts2}{An XTS object that may or may not have indices that fall within
the range of \code{xts1}, but do not appear in \code{xts1}}

\item{agg_function}{Any function that accepts a numeric vector as input and
returns a single number as output; e.g., sum}

\item{...}{Additional parameters to pass to \code{agg_function}}
}
\description{
For two XTS objects \code{xts1} and \code{xts2}, let \code{A}
be any datetime index that appears in \code{index(xts2)}, falls between
\code{min(index(xts1))} and \code{max(index(xts1))}, but does \bold{not}
appear in \code{index(xts1)}. Let \code{B} be the earliest-occurring
datetime index in \code{index(xts1)} that occurs \bold{after} \code{A}.

\code{xts_merge_align_next(xts1, xts2)} will return the left-join
merge of \code{xts1} and \code{xts2} with one additional operation: for
any/all \code{A} and \code{B} that occur for the XTS pair, the values at
at index \code{B} for the columns in the merged XTS coming from \code{xts2}
equal the aggregated value (as specified; e.g., \code{sum}) of the values
of the columns in \code{xts2} in the range \code{A:B}.
}
